+++ src/object_store/src/object/opendal_engine/gcs.rs
@@ -15,18 +15,13 @@
 use opendal::layers::{LoggingLayer, RetryLayer};
 use opendal::services::Gcs;
 use opendal::Operator;
-use risingwave_common::config::ObjectStoreConfig;
 
-use super::{new_http_client, EngineType, OpendalObjectStore};
+use super::{EngineType, OpendalObjectStore};
 use crate::object::ObjectResult;
 
 impl OpendalObjectStore {
     /// create opendal gcs engine.
-    pub fn new_gcs_engine(
-        bucket: String,
-        root: String,
-        object_store_config: ObjectStoreConfig,
-    ) -> ObjectResult<Self> {
+    pub fn new_gcs_engine(bucket: String, root: String) -> ObjectResult<Self> {
         // Create gcs backend builder.
         let mut builder = Gcs::default();
 
@@ -39,9 +34,6 @@ impl OpendalObjectStore {
         if let Ok(cred) = cred {
             builder.credential(&cred);
         }
-
-        let http_client = new_http_client(&object_store_config)?;
-        builder.http_client(http_client);
 
         let op: Operator = Operator::new(builder)?
             .layer(LoggingLayer::default())
