+++ src/object_store/src/object/opendal_engine/opendal_s3.rs
@@ -15,11 +15,12 @@
 use std::time::Duration;
 
 use opendal::layers::{LoggingLayer, RetryLayer};
+use opendal::raw::HttpClient;
 use opendal::services::S3;
 use opendal::Operator;
 use risingwave_common::config::ObjectStoreConfig;
 
-use super::{new_http_client, EngineType, OpendalObjectStore};
+use super::{EngineType, OpendalObjectStore};
 use crate::object::ObjectResult;
 
 impl OpendalObjectStore {
@@ -40,7 +41,7 @@ impl OpendalObjectStore {
             builder.enable_virtual_host_style();
         }
 
-        let http_client = new_http_client(&object_store_config)?;
+        let http_client = Self::new_http_client(&object_store_config)?;
         builder.http_client(http_client);
 
         let op: Operator = Operator::new(builder)?
@@ -62,5 +63,19 @@ impl OpendalObjectStore {
             op,
             engine_type: EngineType::S3,
         })
+    }
+
+    pub fn new_http_client(config: &ObjectStoreConfig) -> ObjectResult<HttpClient> {
+        let mut client_builder = reqwest::ClientBuilder::new();
+
+        if let Some(keepalive_ms) = config.s3.object_store_keepalive_ms.as_ref() {
+            client_builder = client_builder.tcp_keepalive(Duration::from_millis(*keepalive_ms));
+        }
+
+        if let Some(nodelay) = config.s3.object_store_nodelay.as_ref() {
+            client_builder = client_builder.tcp_nodelay(*nodelay);
+        }
+
+        Ok(HttpClient::build(client_builder)?)
     }
 }
